requirements.txt
faster-whisper
edge-tts
google-genai
pydub
python-dotenv
yt-dlp
